+++ source/ChanSort.Api/Model/ChannelList.cs
@@ -12,8 +12,6 @@ namespace ChanSort.Api
     private int insertProgramNr = 1;
     private int duplicateUidCount;
     private int duplicateProgNrCount;
-    public int FavListCount { get; set; }
-
 
     public static List<string> DefaultVisibleColumns { get; set; } = new List<string>(); // initialized by MainForm